refactor: ♻️ Move util functions to appropriate folder
To follow django organizational standards utility functions should be in a utils folder. I've also renamed the files to be the system of record to make the name space more sensible.

Moved utility functions to a new path.
