+++ nautobot_ssot_aristacv/diffsync/tocv/cloudvision.py
@@ -1,7 +1,7 @@
 """DiffSync adapter for Arista CloudVision."""
 from diffsync import DiffSync
 
-import nautobot_ssot_aristacv.diffsync.cvutils as cvutils
+from nautobot_ssot_aristacv.utils import cloudvision
 
 from .models import UserTag
 
@@ -22,16 +22,16 @@ class CloudVision(DiffSync):
 
     def load(self):
         """Load tag data from CloudVision."""
-        user_tags = cvutils.get_tags_by_type()
+        user_tags = cloudvision.get_tags_by_type()
         for tag in user_tags:
             self.tag = UserTag(name=tag["label"], value=tag["value"])
             self.add(self.tag)
 
-        devices = cvutils.get_devices()
+        devices = cloudvision.get_devices()
         for dev in devices:
             hostname = dev["hostname"]
             # Filter device tags to user-defined tags only
-            dev_tags = [tag for tag in cvutils.get_device_tags(device_id=dev["device_id"]) if tag in user_tags]
+            dev_tags = [tag for tag in cloudvision.get_device_tags(device_id=dev["device_id"]) if tag in user_tags]
             for tag in dev_tags:
                 cur_tag = self.get(UserTag, f"{tag['label']}__{tag['value']}")
                 cur_tag.devices.append(hostname)
